Take RuntimeApiCollection out
