+++ node/service/src/service/darwinia.rs
@@ -25,11 +25,6 @@ pub use sc_service::{
 	ChainSpec, Configuration, TFullBackend, TFullClient, TLightBackend, TLightClient,
 };
 
-use crate::client::DarwiniaClient;
-use crate::service::{
-	set_prometheus_registry, FullBackend, FullClient, FullGrandpaBlockImport, FullSelectChain,
-	LightBackend, LightClient,
-};
 pub use darwinia_runtime;
 
 // --- std ---
@@ -63,7 +58,12 @@ use sp_inherents::InherentDataProviders;
 use sp_runtime::traits::BlakeTwo256;
 use sp_trie::PrefixedMemoryDB;
 // --- darwinia ---
-use darwinia_primitives::{AccountId, Balance, Hash, Nonce, OpaqueBlock as Block, Power};
+use crate::client::DarwiniaClient;
+use crate::service::{
+	set_prometheus_registry, FullBackend, FullClient, FullGrandpaBlockImport, FullSelectChain,
+	LightBackend, LightClient, RuntimeApiCollection,
+};
+use darwinia_primitives::OpaqueBlock as Block;
 use darwinia_rpc::darwinia::{
 	BabeDeps, DenyUnsafe, FullDeps, GrandpaDeps, LightDeps, RpcExtension, SubscriptionTaskExecutor,
 };
@@ -73,49 +73,6 @@ native_executor_instance!(
 	darwinia_runtime::api::dispatch,
 	darwinia_runtime::native_version,
 );
-
-/// A set of APIs that darwinia-like runtimes must implement.
-pub trait RuntimeApiCollection:
-	sp_api::ApiExt<Block, Error = sp_blockchain::Error>
-	+ sp_api::Metadata<Block>
-	+ sp_authority_discovery::AuthorityDiscoveryApi<Block>
-	+ sp_block_builder::BlockBuilder<Block>
-	+ sp_consensus_babe::BabeApi<Block>
-	+ sp_finality_grandpa::GrandpaApi<Block>
-	+ sp_offchain::OffchainWorkerApi<Block>
-	+ sp_session::SessionKeys<Block>
-	+ sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
-	+ frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Nonce>
-	+ pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
-	+ darwinia_balances_rpc_runtime_api::BalancesApi<Block, AccountId, Balance>
-	+ darwinia_header_mmr_rpc_runtime_api::HeaderMMRApi<Block, Hash>
-	+ darwinia_staking_rpc_runtime_api::StakingApi<Block, AccountId, Power>
-where
-	<Self as sp_api::ApiExt<Block>>::StateBackend: sp_api::StateBackend<BlakeTwo256>,
-{
-}
-impl<Api> RuntimeApiCollection for Api
-where
-	Api: sp_transaction_pool::runtime_api::TaggedTransactionQueue<Block>
-		+ sp_api::ApiExt<Block, Error = sp_blockchain::Error>
-		+ sp_api::Metadata<Block>
-		+ sp_authority_discovery::AuthorityDiscoveryApi<Block>
-		+ sp_block_builder::BlockBuilder<Block>
-		+ sp_consensus_babe::BabeApi<Block>
-		+ sp_finality_grandpa::GrandpaApi<Block>
-		+ sp_offchain::OffchainWorkerApi<Block>
-		+ sp_session::SessionKeys<Block>
-		+ frame_system_rpc_runtime_api::AccountNonceApi<Block, AccountId, Nonce>
-		+ pallet_transaction_payment_rpc_runtime_api::TransactionPaymentApi<Block, Balance>
-		+ darwinia_balances_rpc_runtime_api::BalancesApi<Block, AccountId, Balance>
-		+ darwinia_header_mmr_rpc_runtime_api::HeaderMMRApi<Block, Hash>
-		+ darwinia_staking_rpc_runtime_api::StakingApi<Block, AccountId, Power>,
-	<Self as sp_api::ApiExt<Block>>::StateBackend: sp_api::StateBackend<BlakeTwo256>,
-{
-}
-
-pub trait RuntimeExtrinsic: codec::Codec + Send + Sync + 'static {}
-impl<E> RuntimeExtrinsic for E where E: codec::Codec + Send + Sync + 'static {}
 
 #[cfg(feature = "full-node")]
 fn new_partial<RuntimeApi, Executor>(
